+++ lib/tools/text_tools/date_parser/date_parsed_widget.dart
@@ -20,19 +20,19 @@ class DateParsedWidget extends StatelessWidget {
           ),
         );
     final tzOffset = date.timeZoneOffset.inHours;
-    final weekday =
-        MaterialLocalizations.of(context).narrowWeekdays[date.weekday];
+    final weekdays = MaterialLocalizations.of(context).narrowWeekdays;
+    String weekday = date.weekday == 7 ? weekdays[0] : weekdays[date.weekday];
     return Wrap(
       spacing: 4,
       runSpacing: 4,
       children: [
-        coloredChip('${date.year}', 'Year'),
-        coloredChip('${date.month}', 'Month'),
-        coloredChip('${date.day}', 'Day'),
-        coloredChip('${date.hour}', 'Hour'),
-        coloredChip('${date.minute}', 'Minute'),
-        coloredChip('${date.second}', 'Second'),
-        coloredChip('${date.millisecond}', 'Millisecond'),
+        coloredChip('${date.year}'.padLeft(4, '0'), 'Year'),
+        coloredChip('${date.month}'.padLeft(2, '0'), 'Month'),
+        coloredChip('${date.day}'.padLeft(2, '0'), 'Day'),
+        coloredChip('${date.hour}'.padLeft(2, '0'), 'Hour'),
+        coloredChip('${date.minute}'.padLeft(2, '0'), 'Minute'),
+        coloredChip('${date.second}'.padLeft(2, '0'), 'Second'),
+        coloredChip('${date.millisecond}'.padLeft(3, '0'), 'Millisecond'),
         coloredChip(weekday, 'Weekday'),
         if (date.isUtc) coloredChip('UTC'),
         coloredChip(date.timeZoneName),